Address Noe's refactoring feedback
